알려진 문제 모두 해결 근데 시간 지날수록 (점이 많아질수록) 속도 느려짐 (fps 작아짐)
그리고 충돌도 추가해야됨

+++ src/app/simulations/state.rs
@@ -77,7 +77,7 @@ impl SimulationSettings {
 pub struct SimulationState {
     pub settings: SimulationSettings,
     pub(crate) pointer: Option<PlotPoint>,
-    pub(crate) time_mul: f64,
+    pub(crate) time_mul: usize,
     pub(crate) time: f64,
     pub(crate) current_step: usize,
     pub(crate) max_step: usize,
@@ -95,7 +95,7 @@ impl Default for SimulationState {
             max_step: 0,
             sim_started: false,
             zoom: 1.0,
-            time_mul: 1.0,
+            time_mul: 1,
         }
     }
 }
